Change the namespace of the Renderer class
